refactor(web): split glossary list and concept detail
